Add forgotten abstract method
Abstract portable cells to allow for portables with custom cell inventories

Add default to ICellWorkbenchItem::isEditable

Move portable storage tiers to a dedicated class
Allows for future add-on instantiation and use in regular cells
