+++ src/main/java/appeng/items/tools/powered/MatterCannonItem.java
@@ -383,11 +383,6 @@ public class MatterCannonItem extends AEBasePoweredItem implements IBasicCellIte
     }
 
     @Override
-    public boolean isEditable(ItemStack is) {
-        return true;
-    }
-
-    @Override
     public IUpgradeInventory getUpgrades(ItemStack is) {
         return UpgradeInventories.forItem(is, 4, this::onUpgradesChanged);
     }
